+++ plugins/announcements/src/components/Router.tsx
@@ -23,6 +23,9 @@ type RouterProps = {
   themeId?: string;
   title?: string;
   subtitle?: string;
+  cardOptions?: {
+    titleLength: number | undefined;
+  };
 };
 
 export const Router = (props: RouterProps) => {